Moved sprite functions to separate header
